+++ src/background/exposed.rs
@@ -187,6 +187,7 @@ pub fn register_invoke_handler(app_builder: Builder<Wry>) -> Builder<Wry> {
         state_get_layouts,
         state_get_weg_items,
         state_get_settings,
+        state_write_settings,
         state_get_specific_apps_configurations,
         state_get_wallpaper,
         state_set_wallpaper,